test: update splits and re-run
